sources/fase1/requirements.txt
requests>=2.31.0
pandas>=1.5.3
numpy>=1.24.3
python-dotenv>=1.0.0
fastapi>=0.95.0
uvicorn>=0.21.1
rpy2>=3.5.14
scipy>=1.10.1
matplotlib>=3.7.1
seaborn>=0.12.2 